ux500-ARM : make backup RAM as an executable area
On the v1.0/ED boards, during the deep sleep resume, it is
required to make the backup RAM area as "executable" in order
manage MMU settings for a single core mode. As a result, the
rom code requires such a condition for aligning SMP strategy
for different per-core MMU configuration. This patch makes
the backup RAM configuration area as executable by adding
a new memory type.

*This change will not be needed for U8500 v2.0 as the
 rom code would (hopefully as discussions are showing) be updated
 with a minor API change*

This patch is a part of patches for the deep sleep feature
for the U8500 and FIDO_IR_ER: 258539

This patch *must* be re-visisted during u8500 v2.0 cut
or mainlining the deep sleep to community

Signed-off-by: Sundar R Iyer <sundar.iyer@stericsson.com>
Acked-By: Biju C Das <biju.das@stericsson.com>
Signed-off-by: Mian Yousaf Kaukab <mian.yousaf.kaukab@stericsson.com>
Change-Id: Id4d8215ca1e9aaf8f327ff3f55f3f9ca1e68aef5
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/2190
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>

Conflicts:
	arch/arm/mm/mmu.c