mmc: block: add eMMC hardware reset support  901ac75c - https://github.com/ChronoMonochrome/android_kernel_ste-3.4/commit/67716327eec7e9d573e7cb2d806545d6f7c1a38d
For cards that support hardware reset (just eMMC), try a reset and
retry before returning an I/O error.  However this is not done for
ECC errors and is never done twice for the same operation type
(READ, WRITE, DISCARD, SECURE DISCARD) until that type of operation
again succeeds.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Chris Ball <cjb@laptop.org>

Conflicts:
	drivers/mmc/card/block.c